moving to a modularized repo structure. Fixed zoom method on desktop. Added NUTS and LAU codes to tooltip using GISCO ID API.
