+++ examples/croatia/index.html
@@ -1,59 +1,61 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Croatian Tourism Grid</title>
-    <link rel="shortcut icon" href="../../assets/images/favicon.ico" />
-    <meta name="viewport" content="width=device-width" />
-  </head>
 
-  <body style="margin: 0">
-    <div id="viz-container"></div>
-  </body>
-  <script src="../../build/gridviz.js"></script>
-  <script>
-    let containerDiv = document.getElementById("viz-container");
-    gridviz
-      .viewer()
-      .container(containerDiv)
-      .title("Tourism Grid of Croatia")
-      .gridData([
-        {
-          url: "../../assets/csv/Croatia/croatia_tourism_grid.csv",
-          cellSize: 1000,
-        },
-      ])
-      .colorField("NIGHTS_TOT") // csv column which you wish to visualize
-      .zoom(500000) //initial camera Z value
-      .nuts2json(true) //loads borders topojson
-      .nuts2jsonCountry("HR") //filter nuts2json
-      .placenamesCountry("HR") //filter placenames
-      .placenameThresholds({
-        //"scale": population
-        450000: 200000,
-        200000: 10000,
-        100000: 5000,
-        10000: 5000,
-        1000: 500,
-      })
-      .center([4755000, 2369000]) //EPSG=3035
-      .height(window.innerHeight - 4)
-      .width(window.innerWidth - 4)
-      .backgroundColor("#000")
-      .borderColor("#ffffff")
-      .colorSchemeName("interpolatePlasma")
-      .colorSchemeSelector(true)
-      .colorScaleSelector(true)
-      .colorFieldSelector(true)
-      .colorScaleSelectorLabel("Scale: ")
-      .colorFieldSelectorLabel("Indicator: ")
-      .legend({
-        width: 350,
-        title: "Total tourist nights per km² (2017)",
-        titleWidth: 150,
-        cells: 10,
-        format: ",.2r",
-      })
-      .build();
-  </script>
+<head>
+  <meta charset="utf-8" />
+  <title>Croatian Tourism Grid</title>
+  <link rel="shortcut icon" href="../../assets/images/favicon.ico" />
+  <meta name="viewport" content="width=device-width" />
+</head>
+
+<body style="margin: 0">
+  <div id="viz-container"></div>
+</body>
+<script src="../../build/gridviz.js"></script>
+<script>
+  let containerDiv = document.getElementById("viz-container");
+  gridviz
+    .viewer()
+    .container(containerDiv)
+    .title("Tourism Grid of Croatia")
+    .gridData([
+      {
+        url: "../../assets/csv/Croatia/croatia_tourism_grid.csv",
+        cellSize: 1000,
+      },
+    ])
+    .colorField("NIGHTS_TOT") // csv column which you wish to visualize
+    .zoom(500000) //initial camera Z value
+    .nuts(true) //loads borders topojson
+    .nutsCountry("HR") //filter nuts
+    .placenamesCountry("HR") //filter placenames
+    .placenameThresholds({
+      //"scale": population
+      450000: 200000,
+      200000: 10000,
+      100000: 5000,
+      10000: 5000,
+      1000: 500,
+    })
+    .center([4755000, 2369000]) //EPSG=3035
+    .height(window.innerHeight - 4)
+    .width(window.innerWidth - 4)
+    .backgroundColor("#000")
+    .lineColor("#ffffff")
+    .colorSchemeName("interpolatePlasma")
+    .colorSchemeSelector(true)
+    .colorScaleSelector(true)
+    .colorFieldSelector(true)
+    .colorScaleSelectorLabel("Scale: ")
+    .colorFieldSelectorLabel("Indicator: ")
+    .legend({
+      width: 350,
+      title: "Total tourist nights per km² (2017)",
+      titleWidth: 150,
+      cells: 10,
+      format: ",.2r",
+    })
+    .build();
+</script>
+
 </html>
