+++ examples/france/population-over-80.html
@@ -26,12 +26,12 @@
         .height(window.innerHeight)
         .width(window.innerWidth)
         .backgroundColor("#000")
-        .borderColor("#ffffff")
+        .lineColor("#ffffff")
         .colorSchemeName("interpolateInferno")
         .colorSchemeSelector(true)
-        .nuts2json(true)
-        .nuts2jsonCountry("FR")
-        .nuts2jsonEPSG(3035)
+        .nuts(true)
+        .nutsCountry("FR")
+        .EPSG(3035)
         .placenamesCountry("FR")
         .placenameThresholds({ //"scale": population
             "1000000": 100000,
