+++ examples/portugal/index.html
@@ -33,16 +33,16 @@
         ])
         .colorField("TOT_P")  // initial csv column to visualize
         .zoom(2000) //initial camera Z value
-        .nuts2json(true) //loads nuts2json topojson boundaries
-        .nutsSimplification("10M") //loads nuts2json topojson boundaries
-        .nuts2jsonCountry("PT") //filter nuts2json
+        .nuts(true) //loads nuts topojson boundaries
+        .nutsSimplification("10M") //loads nuts topojson boundaries
+        .nutsCountry("PT") //filter nuts
         .showPlacenames(true)
         .placenamesCountry("PT") //filter placenames
         .zerosRemoved(3)
         .sourcesHTML("Source: <a href='https://www.efgs.info/data/national/' target='_blank'>Statistics Portugal</a>")
         .center([2677, 1964])
         .backgroundColor("#fcf6c2")
-        .borderColor("#000")
+        .lineColor("#000")
         .colorSchemeName("interpolateRdYlGn") //d3-scale-chromatic
         .reverseColorScheme(true)
         .colorSchemeSelector(true)
