+++ readme.md
@@ -88,7 +88,7 @@ Here's a barebones example that loads a CSV containing population data for a 5 k
 | viewer.height([value])          | int         | window.innerHeight | The height value used for the threejs scene, camera, zoom, labels.                               |
 | viewer.width([value])           | int         | window.innerWidth  | The width value used for the threejs scene, camera, zoom, labels.                                |
 | viewer.backgroundColor([value]) | String      | "#b7b7b7"          | Viewer background colour                                                                         |
-| viewer.borderColor([value])     | String      | "#ffffff"          | Colour used for line geometries                                                                  |
+| viewer.lineColor([value])       | String      | "#ffffff"          | Colour used for line geometries                                                                  |
 | viewer.highlightColor([value])  | String      | "#37f2d6"          | Hex string which is later converted to THREE.Line2 color (hexidecimal).                          |
 | viewer.loadingIcon([value])     | String      | "ring"             | CSS animation used for the loading icon, other options are: ripple \| ring \| ellipsis \|roller. |
 | viewer.title([value])           | String      | null               | Appends a title to the viewer                                                                    |
@@ -107,11 +107,11 @@ Here's a barebones example that loads a CSV containing population data for a 5 k
 | viewer.gridData([value])             | Object {url:string, cellSize: int} | null    | Object containing the URL of the CSV file and its cellSize (this will determine the point size in three.js). A 1 km² grid will have a cellSize of 1000.       |
 | viewer.center([value])               | Array                              | null    | [x,y] coordinates with which the viewer will center the camera.                                                                                               |
 | viewer.addGeoJson([value])           | String                             | null    | Loads a geojson file from the specified URL and adds it's geometries to the viewer. Currently only accepts "polygon" or "multipolygon" geometries             |
-| viewer.nuts2json([value])            | Boolean                            | false   | Show NUTS boundaries using [nuts2json](https://github.com/eurostat/nuts2json)                                                                                 |
-| viewer.nuts2jsonEPSG([value])        | int                                | 3035    | EPSG code of the projection of the nuts2json geometries to request. (available in 3035, 3857, 4258 or 4326)                                                   |
+| viewer.nuts([value])                 | Boolean                            | false   | Show NUTS boundaries using [nuts2json](https://github.com/eurostat/nuts2json)                                                                                 |
+| viewer.EPSG([value])                 | int                                | 3035    | EPSG code of the projection to use for the geometries, placename labels and NUTS boundaries. (NUTS only available in 3035, 3857, 4258 or 4326)                |
 | viewer.nutsLevel([value])            | int                                | 0       | Nuts2json NUTS level                                                                                                                                          |
 | viewer.nutsSimplification([value])   | String                             | "20M"   | Nuts2json simplification level                                                                                                                                |
-| viewer.nuts2jsonCountry([value])     | String                             | null    | Filters nuts2json geometries by country code. Used in order to only load boundaries for a single country.                                                     |  |
+| viewer.nutsCountry([value])          | String                             | null    | Filters nuts2json geometries by country code. Used in order to only load boundaries for a single country.                                                     |  |
 | viewer.showPlacenames([value])       | Boolean                            | false   | Adds placenames to the viewer. Placename-scale thresholds currently defined by the user.                                                                      |
 | viewer.placenamesCountry([value])    | String                             | null    | Filters placenames by country code                                                                                                                            |
 | viewer.placenamesThresholds([value]) | Object                             | null    | Defines population query parameter at certain scale thresholds e.g. {"100000":500} will show placenames with populations of over 5000 at scales up to 100 000 |
