+++ src/css/legend.css
@@ -1,0 +1,45 @@
+/* LEGEND */
+#gridviz-legend {
+    z-index: 9999;
+    position: absolute;
+    bottom: 1%; 
+    left: 1%;
+  
+    padding-left: 15px;
+    padding-right: 15px;
+    padding-top: 5px;
+    padding-bottom: 5px;
+  }
+  
+  /* continuous */
+  .gridviz-legend-svg{
+    font-size:0.8rem;
+    margin-top: 5px;
+  }
+  .gridviz-continuous-legend-title{
+    font-size:0.8rem;
+    font-weight:bold;
+    margin-bottom:10px;
+  }
+  .legend-tick text{
+  /* font-size:1rem */
+  }
+  .gridviz-legend-axis .domain{
+    visibility:hidden;
+  }
+  
+  .gridviz-legend-horizontal {
+    height: 65px;
+    width: 441px;
+    padding-top: 5px;
+  }
+  .gridviz-legend-vertical {
+    height: 321px;
+    width: 140px;
+    padding-top: 5px;
+  }
+  /* legend cells */
+  rect.swatch {
+    stroke: white;
+    stroke-width: 0.5px;
+  }
